gateway/requirements.txt
aiofiles==23.1.0
asgiref==3.7.2
blinker==1.5
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.14
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
opentelemetry-api==1.18.0
opentelemetry-distro==0.39b0
opentelemetry-exporter-zipkin-json==1.18.0
opentelemetry-exporter-zipkin-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-asgi==0.39b0
opentelemetry-instrumentation-aws-lambda==0.39b0
opentelemetry-instrumentation-dbapi==0.39b0
opentelemetry-instrumentation-jinja2==0.39b0
opentelemetry-instrumentation-logging==0.39b0
opentelemetry-instrumentation-requests==0.39b0
opentelemetry-instrumentation-sqlite3==0.39b0
opentelemetry-instrumentation-urllib==0.39b0
opentelemetry-instrumentation-wsgi==0.39b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
priority==2.0.0
protobuf==3.20.3
quart==0.18.4
requests==2.31.0
toml==0.10.2
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
Werkzeug==2.3.6
wrapt==1.15.0
wsproto==1.2.0
zipp==3.15.0
